Refactor bridge/main.go into separate files for better LLM parsing
Split the 4820-line main.go into 11 focused files to improve token
efficiency for LLM-assisted development:

- protocol.go: IPC framing protocol (76 lines)
- types.go: Core types and custom widgets (292 lines)
- widget_creators.go: All Create* widget handlers (1404 lines)
- window.go: Window management (242 lines)
- widget_properties.go: Get/Set property handlers (678 lines)
- widget_state.go: Visibility/state handlers (213 lines)
- containers.go: Container operations (249 lines)
- interactions.go: Testing/interaction handlers (845 lines)
- dialogs.go: Dialog handlers (179 lines)
- misc.go: Utility handlers (445 lines)
- main.go: Entry point and message dispatcher (282 lines)

Benefits:
- Each file can be parsed independently by LLMs
- Better code organization by functional area
- No performance impact (Go compiles at package level)
- Easier maintenance and navigation

Total: 4905 lines (85 lines overhead for package declarations)
Original: 4820 lines in single file
